various improvements

+++ cron/1.characters.nolock.php
@@ -13,7 +13,7 @@ if ($redis->get("zkb:noapi") == "true") exit();
 $esiCorps = new RedisTimeQueue('tqCorpApiESI', 3600);
 $esi = new RedisTimeQueue('tqApiESI', 3600);
 if ($mt == 0 && (date('i') == 22 || $esi->size() < 100)) {
-    Log::log("populating tqApiESI: " . $esi->size());
+    //Log::log("populating tqApiESI: " . $esi->size());
     $esis = $mdb->find("scopes", ['scope' => 'esi-killmails.read_killmails.v1']);
     foreach ($esis as $row) {
         $charID = $row['characterID'];
@@ -38,15 +38,15 @@ while ($minute == date('Hi')) {
                 $mdb->set("scopes", $row, ['corporationID' => $corpID], true);
             }
 
-            //$hasRecent = $mdb->exists("ninetyDays", ['involved.characterID' => $charID]);
-            //if (!$hasRecent && @$row['lastFetch']->sec != 0 && (($charID % 24) != date('H'))) {usleep(100000); continue; }
+            $hasRecent = $mdb->exists("ninetyDays", ['involved.characterID' => $charID]);
+            if (!$hasRecent && @$row['lastFetch']->sec != 0 && (($charID % 24) != date('H'))) continue;
 
             $params = ['row' => $row, 'esi' => $esi];
             $refreshToken = $row['refreshToken'];
             $accessToken = $sso->getAccessToken($refreshToken);
             if (is_array($accessToken) && @$accessToken['error'] == "invalid_grant") {
                 $mdb->remove("scopes", $row);
-                sleep(1);
+                //sleep(1);
                 continue;
             }
